requirements.txt
streamlit==1.28.0
pandas==2.0.0
joblib==1.3.0
mysql-connector-python==8.0.33
bcrypt==4.0.1
scikit-learn==1.3.2
numpy==1.24.3
